+++ packages/sample-app/src/@types/dynamic-plugin-sdk.d.ts
@@ -1,0 +1,6 @@
+import type { SupportedCustomProperties } from '../types';
+
+declare module '@openshift/dynamic-plugin-sdk' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  interface PluginCustomProperties extends SupportedCustomProperties {}
+}