+++ packages/sample-plugin/plugin-metadata.ts
@@ -1,3 +1,4 @@
+import type { SupportedCustomProperties } from '@monorepo/sample-app/src/types';
 import type { PluginBuildMetadata } from '@openshift/dynamic-plugin-sdk-webpack';
 
 const metadata: PluginBuildMetadata = {
@@ -10,8 +11,10 @@ const metadata: PluginBuildMetadata = {
     testComponent: './src/test-component',
   },
   customProperties: {
-    test: true,
-  },
+    sampleApp: {
+      greeting: 'Greetings from sample plugin',
+    },
+  } as SupportedCustomProperties,
 };
 
 export default metadata;