+++ packages/lib-core/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 6.1.0 - TBD
 
-- Update yup to 1.7.1 ([#289])
+- Updated `yup` dependency to `^1.7.1`. ([#289])
+- Allow extending `customProperties` object type in `PluginRuntimeMetadata` ([#290])
 
 ## 6.0.0 - 2026-01-06
 
@@ -95,3 +96,4 @@
 [#280]: https://github.com/openshift/dynamic-plugin-sdk/pull/280
 [#281]: https://github.com/openshift/dynamic-plugin-sdk/pull/281
 [#289]: https://github.com/openshift/dynamic-plugin-sdk/pull/289
+[#290]: https://github.com/openshift/dynamic-plugin-sdk/pull/290